+++ invenio_rdm_records/oaiserver/services/services.py
@@ -18,12 +18,14 @@ from invenio_records_resources.services.base import LinksTemplate
 from invenio_records_resources.services.records.schema import ServiceSchemaWrapper
 from invenio_records_resources.services.uow import unit_of_work
 from marshmallow import ValidationError
+from sqlalchemy import or_
 from sqlalchemy.orm.exc import NoResultFound
 from sqlalchemy.sql import text
 
 from invenio_rdm_records.oaiserver.services.errors import (
     OAIPMHSetDoesNotExistError,
     OAIPMHSetIDDoesNotExistError,
+    OAIPMHSetNotEditable,
     OAIPMHSetSpecAlreadyExistsError,
 )
 from invenio_rdm_records.oaiserver.services.uow import OAISetCommitOp, OAISetDeleteOp
@@ -32,9 +34,10 @@ from invenio_rdm_records.oaiserver.services.uow import OAISetCommitOp, OAISetDel
 class OAIPMHServerService(Service):
     """OAI-PMH service."""
 
-    def __init__(self, config):
+    def __init__(self, config, extra_reserved_prefixes={}):
         """Init service with config."""
         super().__init__(config)
+        self.reserved_prefixes = config.reserved_prefixes.union(extra_reserved_prefixes)
 
     @property
     def schema(self):
@@ -64,12 +67,11 @@ class OAIPMHServerService(Service):
     def _validate_spec(self, spec):
         """Checks the validity of the provided spec."""
         # Reserved for community integration
-        reserved_prefix = "community-"
-        if spec.startswith(reserved_prefix):
+        if spec.startswith(tuple(self.reserved_prefixes)):
             raise ValidationError(
                 _(
-                    "The spec must not start with '{prefix}'.".format(
-                        prefix=reserved_prefix
+                    "The spec must not start with any of the following list '{prefix}'.".format(
+                        prefix=list(self.reserved_prefixes)
                     )
                 ),
                 field_name="spec",
@@ -96,10 +98,10 @@ class OAIPMHServerService(Service):
             context={"identity": identity},
             raise_errors=True,
         )
-
         self._validate_spec(valid_data["spec"])
+        system_created = valid_data["spec"].startswith(tuple(self.reserved_prefixes))
 
-        new_set = OAISet(**valid_data)
+        new_set = OAISet(**valid_data, system_created=system_created)
         existing_set, errors = self._get_one(spec=new_set.spec, raise_error=False)
         if existing_set:
             raise OAIPMHSetSpecAlreadyExistsError(new_set.spec)
@@ -130,12 +132,27 @@ class OAIPMHServerService(Service):
 
         search_params = self._get_search_params(params)
 
-        oai_sets = OAISet.query.order_by(
-            search_params["sort_direction"](text(",".join(search_params["sort"])))
-        ).paginate(
-            page=search_params["page"],
-            per_page=search_params["size"],
-            error_out=False,
+        query_param = search_params["q"]
+        filters = []
+
+        if query_param:
+            filters.extend(
+                [
+                    OAISet.name.ilike(f"%{query_param}%"),
+                    OAISet.spec.ilike(f"%{query_param}%"),
+                ]
+            )
+
+        oai_sets = (
+            OAISet.query.filter(or_(*filters))
+            .order_by(
+                search_params["sort_direction"](text(",".join(search_params["sort"])))
+            )
+            .paginate(
+                page=search_params["page"],
+                per_page=search_params["size"],
+                error_out=False,
+            )
         )
 
         return self.result_list(
@@ -152,6 +169,8 @@ class OAIPMHServerService(Service):
         """Update an OAI set."""
         self.require_permission(identity, "update")
         oai_set, errors = self._get_one(id=id_)
+        if oai_set.system_created:
+            raise OAIPMHSetNotEditable(oai_set.id)
 
         valid_data, errors = self.schema.load(
             data,
@@ -175,6 +194,8 @@ class OAIPMHServerService(Service):
         """Delete an OAI set."""
         self.require_permission(identity, "delete")
         oai_set, errors = self._get_one(id=id_)
+        if oai_set.system_created:
+            raise OAIPMHSetNotEditable(oai_set.id)
         uow.register(OAISetDeleteOp(oai_set))
 
         return True
@@ -231,9 +252,12 @@ class OAIPMHServerService(Service):
         sort = _search_cls.sort_options.get(_sort_name)
         sort_direction = _search_cls.sort_direction_options.get(_sort_direction_name)
 
+        query_params = params.get("q", "")
+
         return {
             "page": page,
             "size": size,
             "sort": sort.get("fields"),
             "sort_direction": sort_direction.get("fn"),
+            "q": query_params,
         }