+++ tests/fixtures/test_cli.py
@@ -8,19 +8,18 @@
 
 """Tests for the CLI."""
 
-from pathlib import Path
 
-import pytest
 from invenio_access.permissions import system_identity
 from invenio_communities import current_communities
 from invenio_communities.communities.records.api import Community
 from invenio_communities.members import Member
 from invenio_requests import current_requests_service
 from invenio_requests.records import Request
-from invenio_vocabularies.contrib.awards.api import Award
-from invenio_vocabularies.contrib.funders.api import Funder
-from invenio_vocabularies.records.api import Vocabulary
 
+from invenio_rdm_records.cli import (
+    create_records_custom_field,
+    custom_field_exists_in_records,
+)
 from invenio_rdm_records.fixtures.demo import create_fake_community, create_fake_record
 from invenio_rdm_records.fixtures.tasks import (
     create_demo_community,
@@ -29,28 +28,13 @@ from invenio_rdm_records.fixtures.tasks import (
     create_demo_record,
     get_authenticated_identity,
 )
-from invenio_rdm_records.fixtures.vocabularies import VocabulariesFixture
 from invenio_rdm_records.proxies import current_rdm_records_service
 from invenio_rdm_records.records import RDMDraft, RDMRecord
 from invenio_rdm_records.requests import CommunitySubmission
 
 
-@pytest.fixture
-def vocabularies():
-    """Load vocabularies."""
-    vocabularies = VocabulariesFixture(
-        system_identity,
-        Path(__file__).parent / "data/vocabularies.yaml",
-        delay=False,
-    )
-    vocabularies.load()
-    Vocabulary.index.refresh()
-    Award.index.refresh()
-    Funder.index.refresh()
-
-
 def test_create_fake_demo_draft_record(
-    app, location, db, es_clear, vocabularies, users
+    app, location, db, search_clear, vocabularies, users
 ):
     """Assert that demo record creation works without failing."""
     user_id = users[0].id
@@ -71,7 +55,9 @@ def test_create_fake_demo_draft_record(
     assert records.total > 0
 
 
-def test_create_fake_demo_communities(app, location, db, es_clear, vocabularies, users):
+def test_create_fake_demo_communities(
+    app, location, db, search_clear, vocabularies, users
+):
     """Assert that demo communities creation works without failing."""
     user_id = users[0].id
 
@@ -84,7 +70,7 @@ def test_create_fake_demo_communities(app, location, db, es_clear, vocabularies,
 
 
 def test_create_fake_demo_inclusion_requests(
-    app, location, db, es_clear, vocabularies, users
+    app, location, db, search_clear, vocabularies, users
 ):
     """Assert that demo inclusion requests creation works without failing."""
     user_id = users[0].id
@@ -104,7 +90,7 @@ def test_create_fake_demo_inclusion_requests(
 
 
 def test_create_fake_demo_invitation_requests(
-    app, location, db, es_clear, vocabularies, users
+    app, location, db, search_clear, vocabularies, users
 ):
     """Assert that demo invitation requests creation works without failing."""
     first_user_id = users[0].id
@@ -124,3 +110,36 @@ def test_create_fake_demo_invitation_requests(
     service = current_communities.service.members
     reqs = service.search_invitations(system_identity, comm["id"])
     assert reqs.total > 0
+
+
+def test_create_records_custom_fields(app, location, db, search_clear, cli_runner):
+    """Assert that custom fields mappings are created for records."""
+    result = cli_runner(create_records_custom_field, "-f", "cern:myfield")
+    assert result.exit_code == 0
+
+    record_mapping_field = list(RDMRecord.index.get_mapping().values())[0]["mappings"][
+        "properties"
+    ]["custom_fields"]
+    draft_mapping_field = list(RDMDraft.index.get_mapping().values())[0]["mappings"][
+        "properties"
+    ]["custom_fields"]
+    expected_value = {
+        "dynamic": "true",
+        "properties": {
+            "cern:myfield": {"type": "text", "fields": {"keyword": {"type": "keyword"}}}
+        },
+    }
+    assert record_mapping_field == expected_value
+    assert draft_mapping_field == expected_value
+
+    # check for existence
+    RDMRecord.index.refresh()
+    RDMDraft.index.refresh()
+
+    result = cli_runner(custom_field_exists_in_records, "-f", "cern:myfield")
+    assert result.exit_code == 0
+    assert "Field cern:myfield exists" in result.output
+
+    result = cli_runner(custom_field_exists_in_records, "-f", "unknownfield")
+    assert result.exit_code == 0
+    assert "Field unknownfield does not exist" in result.output