+++ tests/fixtures/test_fixtures.py
@@ -11,16 +11,21 @@ from pathlib import Path
 import pytest
 from invenio_access.permissions import system_identity
 from invenio_accounts.proxies import current_datastore
+from invenio_communities import current_communities
+from invenio_communities.fixtures.tasks import create_demo_community
 from invenio_pidstore.errors import PIDDoesNotExistError
 from invenio_records_resources.proxies import current_service_registry
 from invenio_vocabularies.proxies import current_service as vocabulary_service
 
+from invenio_rdm_records.fixtures import RecordsFixture, create_demo_record
+from invenio_rdm_records.fixtures.communities import CommunitiesFixture
 from invenio_rdm_records.fixtures.users import UsersFixture
 from invenio_rdm_records.fixtures.vocabularies import (
     GenericVocabularyEntry,
     PrioritizedVocabulariesFixtures,
     VocabularyEntryWithSchemes,
 )
+from invenio_rdm_records.proxies import current_rdm_records
 
 
 @pytest.fixture(scope="module")
@@ -35,7 +40,7 @@ def affiliations_service(app):
     return current_service_registry.get("affiliations")
 
 
-def test_load_languages(app, db, es_clear):
+def test_load_languages(app, db, search_clear):
     id_ = "languages"
     languages = GenericVocabularyEntry(
         Path(__file__).parent / "data",
@@ -49,7 +54,7 @@ def test_load_languages(app, db, es_clear):
     assert item.id == "aae"
 
 
-def test_load_resource_types(app, db, es_clear):
+def test_load_resource_types(app, db, search_clear):
     id_ = "resourcetypes"
     resource_types = GenericVocabularyEntry(
         Path(__file__).parent / "data",
@@ -68,7 +73,7 @@ def test_load_resource_types(app, db, es_clear):
     assert item_dict["props"]["datacite_general"] == "Collection"
 
 
-def test_load_community_types(app, db, es_clear):
+def test_load_community_types(app, db, search_clear):
     id_ = "communitytypes"
     resource_types = GenericVocabularyEntry(
         Path(__file__).parent / "data",
@@ -87,7 +92,7 @@ def test_load_community_types(app, db, es_clear):
     assert item_dict["title"]["en"] == "Organization"
 
 
-def test_loading_paths_traversal(app, db, es_clear, subjects_service):
+def test_loading_paths_traversal(app, db, search_clear, subjects_service):
     dir_ = Path(__file__).parent
     fixtures = PrioritizedVocabulariesFixtures(
         system_identity,
@@ -141,7 +146,7 @@ def filepath_replaced_by(filepath, replacement):
         backup.replace(filepath)
 
 
-def test_reloading_paths_traversal(app, db, es_clear, subjects_service):
+def test_reloading_paths_traversal(app, db, search_clear, subjects_service):
     dir_ = Path(__file__).parent
     fixtures = PrioritizedVocabulariesFixtures(
         system_identity,
@@ -183,13 +188,87 @@ def test_load_users(app, db, admin_role):
     users.load()
 
     # app_data/users.yaml doesn't create an admin@inveniosoftware.org user
-    u1 = current_datastore.find_user(email="admin@inveniosoftware.org")
-    assert u1 is None
-    assert current_datastore.find_user(email="admin@example.com")
-    assert current_datastore.find_user(email="user@example.com")
+    non_existing_user = current_datastore.find_user(email="admin@inveniosoftware.org")
+    admin = current_datastore.find_user(email="admin@example.com")
+    user = current_datastore.find_user(email="user@example.com")
+
+    assert non_existing_user is None
+
+    assert admin
+    assert admin.confirmed_at
+
+    assert user
+    assert user.confirmed_at is None
 
 
-def test_load_affiliations(app, db, admin_role, es_clear, affiliations_service):
+def test_load_communities(app, db, location):
+    dir_ = Path(__file__).parent
+    service = current_communities.service
+    communities = CommunitiesFixture(
+        [dir_ / "app_data", dir_.parent.parent / "invenio_rdm_records/fixtures/data"],
+        "communities.yaml",
+        create_demo_community,
+        dir_ / "app_data" / "img",
+        delay=False,
+    )
+
+    communities.load()
+
+    # Refresh to make changes live
+    service.record_cls.index.refresh()
+
+    community1 = service.search(system_identity, q=f"slug:community1")
+    community2 = service.search(system_identity, q=f"slug:community2")
+
+    assert community1.total == 1
+    assert community2.total == 1
+
+    community1_id = list(community1.hits)[0]["id"]
+    community2_id = list(community2.hits)[0]["id"]
+
+    # Check that community1 is featured
+    community1_featured_entry = service.featured_list(system_identity, community1_id)
+    assert community1_featured_entry.total == 1
+
+    # Check that community2 is not featured
+    community2_featured_entry = service.featured_list(system_identity, community2_id)
+    assert community2_featured_entry.total == 0
+
+    # make sure the right logo was uploaded for community1
+    logo = service.read_logo(system_identity, community1_id)
+    with logo.open_stream("rb") as fs1, open(
+        Path(communities.logo_path) / "community1.png", "rb"
+    ) as fs2:
+        assert fs1.read() == fs2.read()
+
+    # make sure community2 has no logo
+    with pytest.raises(FileNotFoundError):
+        service.read_logo(system_identity, community2_id)
+
+
+def test_load_records(app, db, location, vocabularies):
+    dir_ = Path(__file__).parent
+    service = current_rdm_records.records_service
+    records = RecordsFixture(
+        [dir_ / "app_data", dir_.parent.parent / "invenio_rdm_records/fixtures/data"],
+        "records.yaml",
+        create_demo_record,
+        delay=False,
+    )
+
+    records.load()
+
+    # Refresh to make changes live
+    service.record_cls.index.refresh()
+
+    record1 = service.search(system_identity, q=f"Record1")
+    record2 = service.search(system_identity, q=f"Record2")
+
+    assert record1.total == 1
+    assert record2.total == 1
+
+
+def test_load_affiliations(app, db, admin_role, search_clear, affiliations_service):
     dir_ = Path(__file__).parent
     affiliations = VocabularyEntryWithSchemes(
         "affiliations",