+++ tests/resources/test_resources_pids.py
@@ -40,7 +40,7 @@ def publish_record(client, record, headers):
 
 
 def test_external_doi_cleanup(
-    running_app, client, minimal_record, headers, es_clear, uploader
+    running_app, client, minimal_record, headers, search_clear, uploader
 ):
     """Tests for issue #845."""
     client = uploader.login(client)
@@ -77,7 +77,7 @@ def test_external_doi_cleanup(
 
 
 def test_external_doi_duplicate_detection(
-    running_app, client, minimal_record, headers, es_clear, uploader
+    running_app, client, minimal_record, headers, search_clear, uploader
 ):
     """Tests for issue #845."""
     client = uploader.login(client)
@@ -112,7 +112,7 @@ def test_external_doi_duplicate_detection(
 
 
 def test_external_doi_blocked_prefix(
-    running_app, client, minimal_record, headers, es_clear, uploader
+    running_app, client, minimal_record, headers, search_clear, uploader
 ):
     """Tests for issue #847."""
     client = uploader.login(client)
@@ -134,7 +134,7 @@ def test_external_doi_blocked_prefix(
 
 
 def test_external_doi_required(
-    running_app, client, minimal_record, headers, es_clear, uploader
+    running_app, client, minimal_record, headers, search_clear, uploader
 ):
     """Tests for issue #847."""
     client = uploader.login(client)
@@ -147,3 +147,23 @@ def test_external_doi_required(
         {"field": "pids.doi", "messages": ["Missing DOI for required field."]}
     ]
     assert draft.json["pids"] == {"doi": {"provider": "external", "identifier": ""}}
+
+
+def test_pids_publish_validation_error(
+    running_app, client, minimal_record, headers, search_clear, uploader
+):
+    """Ensure that errors raised by Pids component at publish time serialize well."""
+    client = uploader.login(client)
+    del minimal_record["metadata"]["publisher"]
+    draft = client.post("/records", headers=headers, json=minimal_record)
+
+    record = client.post(link(draft.json["links"]["publish"]), headers=headers)
+
+    assert record.status_code == 400
+    expected = [
+        {
+            "field": "metadata.publisher",
+            "messages": ["Missing publisher field required for DOI registration."],
+        }
+    ]
+    assert expected == record.json["errors"]