+++ CHANGES.rst
@@ -1,5 +1,5 @@
 ..
-    Copyright (C) 2019 CERN.
+    Copyright (C) 2019-2023 CERN.
     Copyright (C) 2019 Northwestern University.
 
 
@@ -10,6 +10,53 @@
 Changes
 =======
 
-Version 1.0.0 (released TBD)
+Version 1.3.2 (released 2023-01-30)
+
+- records: remove double permission check on community records search
+
+Version 1.3.1 (released 2023-01-23)
+
+- Add feature flag for archive download endpoint on record and draft resources
+
+Version 1.3.0 (released 2023-01-20)
+
+- add mechanism to validate a record based on each PID provider
+- fix demo records creation adding missing search index prefix on index refresh
+- Fix response status when searching for records of a non-existing community
+- remove validation on DOI discard action
+- skips PIDs modification when no data is passed
+
+Version 1.2.1 (released 2022-12-01)
+
+- Add identity to links template expand method.
+
+Version 1.2.0 (released 2022-11-29)
+
+- add records fixtures
+
+Version 1.1.0 (released 2022-11-25)
+
+- use communities v4
+- use Axios centralized configuration
+- add i18n translations
+- refactor OAI sets view
+
+Version 1.0.3 (released 2022-11-16)
+
+- add draft indexer in registry
+
+Version 1.0.2 (released 2022-11-15)
+
+- fix service_id config values
+- sanitize html in additional descriptions instead of stripping html
+
+Version 1.0.1 (released 2022-11-04)
+
+- upgrade invenio-vocabularies
+- upgrade invenio-drafts-resources
+- add dynamic formats to administration of oai sets
+- add RO-Crate serializer
+
+Version 1.0.0
 
 - Initial public release.